requirements.txt
# Stock Manager Application - Python Requirements
# No external dependencies required - uses only Python standard library
# Requires Python 3.6 or higher

# The application uses only built-in modules:
# - tkinter (GUI)
# - json (data files)
# - csv (transactions storage)
# - datetime (timestamps)
# - pathlib (file paths)

# Barcode generation utilities
python-barcode>=0.15.1
Pillow>=10.0.0
